+++ Gobchat.App/Module/Updater/NAULocalFileFeedReader.cs
@@ -34,9 +34,10 @@ namespace Gobchat.Core.Module.Updater
             var updateContent = GetContent(feed);
             var appContent = GetContent(GobchatApplicationContext.ApplicationLocation);
 
-            foreach (var file in appContent)
-                if (!updateContent.Contains(file))
-                    tasks.Add(new DeleteTask(Path.Combine(GobchatApplicationContext.ApplicationLocation, file)));
+            // For some reason this tasks crashes the update, seems like I need to fork that project and check out what's going wrong
+            // foreach (var file in appContent)
+            //     if (!updateContent.Contains(file))
+            //         tasks.Add(new DeleteTask(Path.Combine(GobchatApplicationContext.ApplicationLocation, file)));
 
             var files = System.IO.Directory.EnumerateFiles(feed, "*", System.IO.SearchOption.AllDirectories)
                  .GroupBy(s => System.IO.Path.GetDirectoryName(s));